% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{GenDataStrat}
\alias{GenDataStrat}
\title{Generate Data per Stratum}
\usage{
GenDataStrat(n, censor_rate, event_rate, strata, tau)
}
\arguments{
\item{n}{Sample size.}

\item{censor_rate}{Censoring rate.}

\item{event_rate}{Event rate.}

\item{strata}{Stratum label.}

\item{tau}{Truncation time.}
}
\description{
Generate Data per Stratum
}
