% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{GenDataStratum}
\alias{GenDataStratum}
\title{Generate Data per Stratum}
\usage{
GenDataStratum(n, censor_rate, event_rate, stratum, tau)
}
\arguments{
\item{n}{Sample size.}

\item{censor_rate}{Censoring rate.}

\item{event_rate}{Event rate.}

\item{stratum}{Stratum label.}

\item{tau}{Truncation time.}
}
\description{
Generate Data per Stratum
}
