% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargKM.R
\name{MargKM}
\alias{MargKM}
\title{Marginal Kaplan-Meier Curve}
\usage{
MargKM(status, strata, time)
}
\arguments{
\item{status}{Event status.}

\item{strata}{Strata.}

\item{time}{Event time.}
}
\value{
Tabulate marginal KM curve.
}
\description{
Marginal Kaplan-Meier Curve
}
