% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargRMST.R
\name{MargStats}
\alias{MargStats}
\title{Calculate Marginal Statistics}
\usage{
MargStats(est, se, weight, alpha)
}
\arguments{
\item{est}{Statistic value.}

\item{se}{Standard error.}

\item{weight}{Statistic weight.}

\item{alpha}{Type I error.}
}
\value{
Data.frame containing:
\itemize{
  \item Marginalized estimate 'Est' and standard error 'SE'.
  \item Lower 'L' and upper 'U' confidence bounds.
}
}
\description{
Calculate the marginalized RMST and RMTL by taking a weighted combination
of the per-stratum estimates.
}
