% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargSurv.R
\name{OddsRatio}
\alias{OddsRatio}
\title{Calculate Rate Odds Ratio}
\usage{
OddsRatio(alpha = 0.05, rates)
}
\arguments{
\item{alpha}{Type I error.}

\item{rates}{Data.frame containing (strata, arm, rate, se).}
}
\value{
Per-stratum rate odds ratio 'or' and log standard error 'log_se'.
}
\description{
Calculate Rate Odds Ratio
}
