% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\name{PrepCounts}
\alias{PrepCounts}
\title{Stratum Counts}
\usage{
PrepCounts(data, weights)
}
\arguments{
\item{data}{Containing (arm, strata).}

\item{weights}{Weights data.frame as returned by \code{\link{PrepWeights}}.}
}
\value{
Data.frame containing per-stratum counts in each arm.
}
\description{
Stratum Counts
}
