% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\name{PrepWeights}
\alias{PrepWeights}
\title{Calculate Stratum Weights}
\usage{
PrepWeights(data, weight)
}
\arguments{
\item{data}{Containing (time, status, arm, strata).}

\item{weight}{Input weights.}
}
\value{
Data.frame containing 'strata' and 'weight'.
}
\description{
Calculate Stratum Weights
}
