% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.stratRMST}
\alias{print.stratRMST}
\title{Print Method for Stratified RMST Object.}
\usage{
\method{print}{stratRMST}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stratRMST}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{stratRMST}.
}
