% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.stratSurv}
\alias{print.stratSurv}
\title{Print Method for Stratified RMST Object.}
\usage{
\method{print}{stratSurv}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stratSurv}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{stratSurv}.
}
