% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargSurv.R
\name{RateDiff}
\alias{RateDiff}
\title{Calculate Rate Difference.}
\usage{
RateDiff(alpha = 0.05, rates)
}
\arguments{
\item{alpha}{Type I error level.}

\item{rates}{Data.frame containing (strata, arm, rate, se).}
}
\value{
Per-stratum rate differences and standard error.
}
\description{
Calculate Rate Difference.
}
