% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample-data.R
\docType{data}
\name{strat_data}
\alias{strat_data}
\title{Sample Stratified Survival Data}
\format{
A data.frame containing three fields:
\describe{
  \item{arm}{Treatment arm, 0 for reference, 1 for treatment.}
  \item{time}{Observation time between 0 and 21.}
  \item{status}{Status indicator, 0 for censoring, 1 for an event.}
  \item{stratum}{Integer coded as 0, 1, 2.}
}
}
\usage{
data(strat_data)
}
\description{
Synthetic time to event data for 600 patients divided into 3x 200 patient strata.
Within each stratum, there are 100 patients in the treatment arm, and 100 patients
in the reference arm.
}
\keyword{datasets}
