% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargMeans.R
\name{StratMeans}
\alias{StratMeans}
\title{Stratified Mean Analysis}
\usage{
StratMeans(arm, y, strata = NULL, alpha = 0.05, weights = NULL)
}
\arguments{
\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{y}{Continuous outcome.}

\item{strata}{Stratification factor.}

\item{alpha}{Type I error.}

\item{weights}{Per-stratum weights, for sorted strata.}
}
\description{
Calculates the difference and ratio of means. The marginal 
statistics are obtained by taking a weighted combination of the stratum-
specific statistics, where the weights are proportional to the stratum
sizes.
}
