% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargSurv.R
\name{StratRate}
\alias{StratRate}
\title{Stratified RMST Analysis}
\usage{
StratRate(time, status, arm, strata = NULL, tau, alpha = 0.05, weights = NULL)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{strata}{Stratification factor.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{weights}{Per-stratum weights, for sorted strata.}
}
\description{
Calculates the difference and ratio of marginal restricted mean survival 
times (RMSTs) and restricted mean times lost (RMTLs). The marginal 
statistics are obtained by taking a weighted combination of the stratum-
specific statistics, where the weights are proportional to the stratum
sizes.
}
\examples{
# Arm 1.
data1 <- GenData(
  n = 100,
  event_rates = c(0.5, 1),
  tau = 5
)
data1$arm <- 1

# Arm 0.
data0 <- GenData(
  n = 100,
  event_rates = c(1, 1),
  tau = 5
)
data0$arm <- 0

# Overall data set.
data <- rbind(data1, data0)

StratRate(
  time = data$time,
  status = data$status,
  arm = data$arm,
  strata = data$strata,
  tau = 2
)
}
