% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargRMST.R
\name{StratRMST}
\alias{StratRMST}
\title{Stratified RMST Analysis}
\usage{
StratRMST(time, status, arm, strata = NULL, tau, alpha = 0.05, weight = NULL)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{strata}{Stratification factor.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}

\item{weight}{Per-stratum weights, for sorted strata.}
}
\description{
Calculates the difference and ratio of marginal restricted mean survival 
times (RMSTs) and restricted mean times lost (RMTLs). The marginal 
statistics are obtained by taking a weighted combination of the stratum-
specific statistics, where the weights are proportional to the stratum
sizes.
}
\examples{
data(strat_data)
StratRMST(
  time = strat_data$time,
  status = strat_data$status,
  arm = strat_data$arm,
  strata = strat_data$stratum,
  tau = 18
)
}
