% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{stratSurv-class}
\alias{stratSurv-class}
\title{Stratified RMST Object}
\description{
Defines the object class returned by \code{\link{StratRMST}} and
\code{\link{StratRate}}.
}
\section{Slots}{

\describe{
\item{\code{Stratified}}{Stats per arm and stratum.}

\item{\code{Marginal}}{Stats per arm, marginalized over strata.}

\item{\code{Contrasts}}{Contrasts of the marginal statistics.}

\item{\code{Weights}}{Per-stratum weights.}
}}

