% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargRMST.R
\name{StratumRMST}
\alias{StratumRMST}
\title{Calculate Per-Stratum Statistics}
\usage{
StratumRMST(time, status, arm, tau)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{tau}{Truncation time.}
}
\value{
Data.frame containing:
\itemize{
  \item 'Stratum' and 'Arm'.
  \item Truncation time 'Tau'.
  \item Statistic 'Stat' and estimate 'Est'.
  \item Standard error 'SE'.
  \item Lower 'L' and upper 'U'.
}
}
\description{
Calculate the per-stratum and per-arm RMST and RMTL.
}
