% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargKM.R
\name{TabulateKM}
\alias{TabulateKM}
\title{Tabulate Kaplan-Meier}
\usage{
TabulateKM(status, time)
}
\arguments{
\item{status}{Event status.}

\item{time}{Event time.}
}
\value{
Data.frame.
}
\description{
Tabulate Kaplan-Meier
}
