% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,stratSurv-method}
\alias{show,stratSurv-method}
\title{Show Method for Stratified RMST Object}
\usage{
\S4method{show}{stratSurv}(object)
}
\arguments{
\item{object}{An object of class \code{stratSurv}.}
}
\description{
Show Method for Stratified RMST Object
}
