% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{GenData}
\alias{GenData}
\title{Generate Data}
\usage{
GenData(n, censor_rates = 0.2, event_rates = 1, tau = 10)
}
\arguments{
\item{n}{Per-strata sample sizes.}

\item{censor_rates}{Per-strata censoring rates.}

\item{event_rates}{Per-strata event rates. Determines the
number of strata.}

\item{tau}{Truncation time.}
}
\description{
Generate Data
}
