% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargSurv.R
\name{SurvRate}
\alias{SurvRate}
\title{Kaplan-Meier Survival Rate}
\usage{
SurvRate(time, status, tau, alpha = 0.05)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{tau}{Truncation time.}

\item{alpha}{Type I error.}
}
\value{
Data.frame containing:
\itemize{
  \item Truncation time 'tau'.
  \item Survival 'rate'
  \item Standard error 'se'.
}
}
\description{
Calculate the survival rate at a given time point tau.
}
